DS-95, review edits
