+++ next.config.ts
@@ -2,4 +2,4 @@ module.exports = {
   images: {
     unoptimized: true,
   },
-};
+}
