+++ package.json
@@ -9,7 +9,6 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "lint": "eslint --ext .ts,.tsx ./src",
     "next": "15.1.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
