+++ src/components/EventCard/EventCard/EventTags.module.scss
@@ -1,6 +1,10 @@
-.cardTags {
+.tags {
   margin-top: 0.875em;
   display: flex;
   gap: 0.5em;
   flex-wrap: wrap;
+}
+
+.tag {
+  cursor: default;
 }
