--- src/components/event-tags/types/event-tags.ts
@@ -1,4 +1,0 @@
-export interface IEventTagsPresentation {
-  visibleTags: string[];
-  hiddenTagsNumber: number;
-}
