--- src/types/conf-event.ts
@@ -1,8 +1,0 @@
-export interface IConfEventModel {
-  imgLink?: string;
-  title: string;
-  tags?: string[];
-  location: string;
-  dateStart: Date;
-  dateFinish: Date;
-}
